+++ src/styles/index.css
@@ -28,4 +28,3 @@
 @import 'basscss-position';
 @import 'basscss-border';
 @import 'basscss-hide';
-
